backend/pyproject.toml
[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
python_files = ["test_*.py"] 
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "3.11.7"
dependencies = [
    "annotated-types (==0.7.0)",
    "anyio (==4.9.0)",
    "certifi (==2025.4.26)",
    "cffi (==1.17.1)",
    "click (==8.1.8)",
    "coverage (==7.8.0)",
    "cryptography (==44.0.3)",
    "ecdsa (==0.19.1)",
    "fastapi (==0.109.2)",
    "h11 (==0.16.0)",
    "httpcore (==1.0.9)",
    "httpx (==0.26.0)",
    "idna (==3.10)",
    "iniconfig (==2.1.0)",
    "packaging (==25.0)",
    "pluggy (==1.5.0)",
    "pyasn1 (==0.6.1)",
    "pycparser (==2.22)",
    "pydantic (==2.11.4)",
    "pydantic-settings (==2.9.1)",
    "pydantic-core (==2.33.2)",
    "pyjwt (==2.8.0)",
    "pypdf2 (==3.0.1)",
    "pytest (==7.4.4)",
    "pytest-asyncio (==0.23.5)",
    "pytest-cov (==4.1.0)",
    "python-dotenv (==1.1.0)",
    "python-jose (==3.3.0)",
    "python-multipart (==0.0.9)",
    "rsa (==4.9.1)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "sqlalchemy (==2.0.27)",
    "starlette (==0.36.3)",
    "typing-inspection (==0.4.0)",
    "typing-extensions (==4.13.2)",
    "uvicorn (==0.27.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app.main:run"
